Try to render using the dre_pdf file new documents
Starting from 2014-09-18 we have to try to extract the
plain text fro the two columns PDF. This is a delicate
operation, prone to errors.

Another trouble is that we have a few documents in the
future (the old site had several such entries with bad
dates).

Because of this we only try to get the plain text from
the PDFs on documents that are both from the new  site
(timestamp > 2014-09-18) and have dates greater than
2014-09-18.