Statistics gathering app - copy log to a database
This app uses a different database, the name of the database definition
is 'stats' by default, this can be changed using the STATSAPP_DB
setting directive (it should be changed if the name collides with an existing
db def).

The apache log path should be defined using the APACHE_LOG setting
directive. Finally, the log lines are sent to the db using a
transaction, the number of lines sent in each transaction can be set
using the PAGE_SIZE setting directive.

Do not forget, to migrate the new database it must be explicitly done
using the migrate command:

    ./manage.py migrate --database=stats