Better handling of non indexable docs
This is not the final interactionn. Now, when a non indexable document
pops out in a search result it will not be displayed but the result page
will be one result sort. There are cases where an empty page is
displayed. This is not ideal.

To solve this the database has to be reindexed and a flag added to each
document. That way we can search for documents excluding the non
indexable ones.